refactor: deployment utils
